+++ src/main/java/more_rpg_loot/client/entity/renderers/FrostMonarchServantEntityRenderer.java
@@ -14,12 +14,12 @@ import static more_rpg_loot.RPGLoot.MOD_ID;
 
 @Environment(EnvType.CLIENT)
 public class FrostMonarchServantEntityRenderer extends SkeletonEntityRenderer {
-    private static final Identifier TEXTURE =
-            Identifier.of(MOD_ID, "textures/entity/mobs/frostmonarch_servant.png");
+    private static final Identifier TEXTURE = Identifier.of(MOD_ID, "textures/entity/mobs/frostmonarch_servant.png");
+    private static final Identifier EYES = Identifier.of(MOD_ID, "textures/entity/mobs/frostmonarch_servant_overlay.png");
 
     public FrostMonarchServantEntityRenderer(EntityRendererFactory.Context context) {
         super(context, EntityModelLayers.SKELETON, EntityModelLayers.STRAY_INNER_ARMOR, EntityModelLayers.STRAY_OUTER_ARMOR);
-        this.addFeature(new FrostMonarchServantOverlayFeatureRenderer<>(this, context.getModelLoader()));
+        this.addFeature(new FrostMonarchServantOverlayFeatureRenderer<>(this, context.getModelLoader(), EntityModelLayers.STRAY_OUTER, EYES));
     }
 
     public Identifier getTexture(AbstractSkeletonEntity abstractSkeletonEntity) {