+++ src/main/java/more_rpg_loot/client/entity/renderers/FrosthauntEntityRenderer.java
@@ -1,8 +1,8 @@
 package more_rpg_loot.client.entity.renderers;
 
 
+import more_rpg_loot.client.entity.renderers.feature.FrostMonarchServantOverlayFeatureRenderer;
 import more_rpg_loot.client.entity.renderers.feature.FrosthauntEyesFeatureRenderer;
-import more_rpg_loot.client.entity.renderers.feature.FrosthauntOverlayFeatureRenderer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.render.entity.EntityRendererFactory;
@@ -16,13 +16,15 @@ import static more_rpg_loot.RPGLoot.MOD_ID;
 
 @Environment(EnvType.CLIENT)
 public class FrosthauntEntityRenderer extends SkeletonEntityRenderer {
-    private static final Identifier TEXTURE =
-            Identifier.of(MOD_ID, "textures/entity/mobs/frosthaunt.png");
+    private static final Identifier TEXTURE = Identifier.of(MOD_ID, "textures/entity/mobs/frosthaunt.png");
+    private static final Identifier EYES = Identifier.of(MOD_ID, "textures/entity/mobs/frosthaunt_eyes.png");
+    private static final Identifier OVERLAY = Identifier.of(MOD_ID, "textures/entity/mobs/frosthaunt_overlay.png");
 
     public FrosthauntEntityRenderer(EntityRendererFactory.Context context) {
         super(context, EntityModelLayers.SKELETON, EntityModelLayers.STRAY_INNER_ARMOR, EntityModelLayers.STRAY_OUTER_ARMOR);
         this.addFeature(new FrosthauntEyesFeatureRenderer<>(this));
-        this.addFeature(new FrosthauntOverlayFeatureRenderer<>(this, context.getModelLoader()));
+        this.addFeature(new FrostMonarchServantOverlayFeatureRenderer<>(this, context.getModelLoader(), EntityModelLayers.STRAY_OUTER, EYES));
+        this.addFeature(new FrostMonarchServantOverlayFeatureRenderer<>(this, context.getModelLoader(), EntityModelLayers.STRAY_OUTER, OVERLAY));
     }
 
     public Identifier getTexture(AbstractSkeletonEntity abstractSkeletonEntity) {