+++ src/main/java/more_rpg_loot/compat/spell_engine/RPGSeriesLootInjectionLNE.java
@@ -1,7 +1,8 @@
 package more_rpg_loot.compat.spell_engine;
 
+
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
-import net.fabricmc.fabric.api.loot.v2.LootTableEvents;
+import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
 import net.spell_engine.rpg_series.loot.LootConfig;
 import net.spell_engine.rpg_series.loot.LootHelper;
 import net.tinyconfig.ConfigManager;
@@ -9,12 +10,11 @@ import net.tinyconfig.ConfigManager;
 import java.util.HashMap;
 
 import static more_rpg_loot.RPGLoot.MOD_ID;
-import static more_rpg_loot.compat.spell_engine.Default.itemLootConfig;
 
 public class RPGSeriesLootInjectionLNE {
 
     public static ConfigManager<LootConfig> lootEquipmentConfig = new ConfigManager<>
-            ("loot_equipment", itemLootConfig)
+            ("loot_equipment", Default.itemLootConfig)
             .builder()
             .setDirectory(MOD_ID)
             .sanitize(true)
@@ -25,7 +25,11 @@ public class RPGSeriesLootInjectionLNE {
         lootEquipmentConfig.refresh();
         LootHelper.TAG_CACHE.refresh();
         LootTableEvents.MODIFY.register((key, tableBuilder, source, registries) -> {
-            LootHelper.configureV2(registries, key.getValue(), tableBuilder, lootEquipmentConfig.value, new HashMap<>());
+            LootHelper.configureV2(registries,
+                    key.getValue(),
+                    tableBuilder,
+                    lootEquipmentConfig.value,
+                    new HashMap<>());
         });
         ServerLifecycleEvents.SERVER_STARTED.register((server) -> {
             LootHelper.updateTagCache(lootEquipmentConfig.value);