+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ### 1.0.0 - 1.21.1
-- Official 1.21 & 1.21.1 Release!
+# Official 1.21 & 1.21.1 Release!
+
+# CHANGES
+- Added Status Effects for each Boosting Drink & Bowl